MyProjFolder/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
beautifulsoup4==4.12.3
brotli==1.1.0
ConfigParser==7.1.0
cryptography==43.0.3
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
fitz==0.0.1.dev2
HTMLParser==0.0.2
importlib_metadata==8.5.0
ipaddr==2.2.0
keyring==25.5.0
lockfile==0.12.2
lxml==5.3.0
mypy_extensions==1.0.0
numpy==2.1.3
ordereddict==1.1
pandas==2.2.3
protobuf==5.28.3
pyOpenSSL==24.2.1
thread==2.0.5
typing_extensions==4.12.2
zipp==3.20.2
